+++ client/src/components/EditWrite.jsx
@@ -41,6 +41,19 @@ const EditWrite = ({writeMode, page, columns, data, setData, open, setOpen, rowD
         setData(tableData)
     }
 
+    const checkValue = () => {
+        if (!value.content) {
+            window.alert('내용이 작성되지 않았습니다.')
+            return false
+        }
+        if (!value.title) {
+            window.alert('제목이 작성되지 않았습니다.')
+            return false
+        }
+        if (!window.confirm('정말로 저장하시겠습니다.?')) return false
+        return true
+    }
+
     const deleteData = () => {
         let tableData = []
         data.map((prev) => (prev._id!==value._id&&tableData.push(prev)))
@@ -64,20 +77,22 @@ const EditWrite = ({writeMode, page, columns, data, setData, open, setOpen, rowD
     }
 
     const handleUpdate = async () => {
+        const valueStatus = checkValue()
+        if (!valueStatus) return 
+
         let url 
         if (writeMode) {
             url = '/api/' + page + '/write'
         } else {
             url = '/api/' + page + '/update'
         }
-        if(!window.confirm('정말로 저장하시겠습니다.?')) return
         try {
             const res = await axios.post(url, value)
             setValue(res.data)
             if (writeMode) {insertData()}
             else {updateData()}
         } catch (err) {
-            console.log(url, err)
+            window.alert(err)
         }
         handleClose()
     }
@@ -89,7 +104,7 @@ const EditWrite = ({writeMode, page, columns, data, setData, open, setOpen, rowD
             await axios.post(url, value)
             deleteData()
         } catch (err) {
-            console.log(url, err)
+            window.alert(err)
         }
         handleClose()
     }